adding backend, bug fixes

+++ agent/python/dockerhelper.py
@@ -58,7 +58,8 @@ def runImage(image, cpu_shares, mem_limit, disk_limit, network, ports, environme
 def killContainer(taskID):
     if taskID in containers:
         container = containers[taskID]
-        container.kill()
+        if container.status == "running"
+            container.kill()
 
 def getContainerStatus(taskID):
     container = containers[taskID]
